feat (S11P31D102-231) : Experience 추가 &

See merge request s11-final/S11P31D102!95

+++ backend/runIt/src/main/java/com/ssafy/runit/domain/record/dto/response/RecordTodayResponse.java
@@ -5,10 +5,10 @@ import lombok.Builder;
 @Builder
 public record RecordTodayResponse(
         Double distance,
-        Integer time,
+        Long time,
         Integer pace
 ) {
-    public static RecordTodayResponse fromEntity(Double dis, Integer time, Integer pace){
+    public static RecordTodayResponse fromEntity(Double dis, Long time, Integer pace){
         return RecordTodayResponse.builder()
                 .pace(pace)
                 .time(time)
